requirements.txt

Flask
python-dotenv
Markdown
